requirements.txt
amqp==5.0.3
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.5
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-debug-toolbar==3.2
django-filter==2.4.0
django-redis==4.12.1
django-rest-swagger==2.2.0
django-timezone-field==4.1.1
djangorestframework==3.12.2
drf-yasg==1.20.0
Faker==5.6.5
idna==2.10
importlib-metadata==3.4.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
kombu==5.0.2
Markdown==3.3.3
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.8
prompt-toolkit==3.0.11
psycopg2==2.7.3.2
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.5
redis==3.5.3
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
signals==0.0.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
uWSGI==2.0.19.1
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.0
